% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{data.frame-methods}
\alias{as.data.frame.errorcalc}
\alias{data.frame-methods}
\alias{rbind.errorcalc}
\title{Create or bind error objects}
\usage{
\method{as.data.frame}{errorcalc}(x, ...)

rbind.errorcalc(...)
}
\arguments{
\item{x}{an error or error_list object}

\item{...}{additional arguments to \code{data.frame}}

\item{labels}{labels to use instead of \code{names(x)}}

\item{label_name}{name of the label column}
}
\value{
a \code{data.frame}
}
\description{
\code{as.data.frame} converts an error object to a data.frame and \code{rbind}
vertically concatenates two error objects
}
\details{
For error_list objects, \code{as.data.frame} will be called for each list
element and the resulting \code{data.frame}s will be concatentated row-wise.
An addition column labelling the elements will be appended.
}

