% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorstat.R
\docType{package}
\name{errorstat-package}
\alias{errorstat}
\alias{errorstat-package}
\title{MSE & bias computations}
\description{
Computation of MSE, bias and (potentially) other error statistics.
}
\details{
Optionally, these computations can be binned by the true parameter
value.
}

