% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{errorstat}
\alias{errorstat}
\title{Compute one or more error summaries}
\usage{
errorstat(truth, estimate, uppers = NULL, mse = TRUE, bias = TRUE, ...)
}
\arguments{
\item{truth}{true values of the parameter}

\item{estimate}{estimated values of the parameter}

\item{uppers}{upper bounds of each interval}

\item{mse}{should the MSE be computed?}

\item{bias}{should the bias be computed?}

\item{...}{additional error summaries.}
}
\value{
A data.frame containing the number of elements in each interval
and the values of each error summary.
}
\description{
\code{errorstat} computes one or more error summaries (e.g., MSE, bias)
in one or more bins defined by \code{uppers}
}
\details{
\code{errorstat} is primarily a utility for determine how MSE and bias
change with the level of a true parameter. It allows the user to partition
the range true values into disjoint intervals and compute the error
summaries for each of these intervals.

Optionally, the user can define additional error summaries through the
dots argument. Each summary should be a function of two variables. The
first of these variables is the true parameter value, and the second is
the estimated parameter value.
}
\examples{
\dontrun{
library(eRm)
truth <- list(trait = rnorm(100), thresh = rnorm(20))
resp <- evalq(sim.rasch(trait, thresh), truth)
ermfit <- RM(resp)
ppfit <- person.parameter(ermfit)
errorstat(truth, list(trait = coef(ppfit), thresh = -coef(ermfit)), -2:2)
}
}
\author{
Matthew Zeigenfuse
}

