% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{flatten}
\alias{as.data.frame.errorcalc}
\alias{flatten}
\title{"Flatten" error objects}
\usage{
flatten(x, ..., labels = NULL)

\method{as.data.frame}{errorcalc}(x, ...)
}
\arguments{
\item{x}{an errorcalc object of a list of errorcalc object}

\item{...}{optionally, addition errorcalc objects}

\item{labels}{optional column labels for the resulting data.frame}
}
\value{
a \code{data.frame}
}
\description{
Flatten nested error objects into data.
}
\details{
The purpose of this function is to convert lists of error statistics into
a single data.frame, so that they can be easily plotted or summarized. The
result data.frame contains the various error computations stacked vertically
with one or more additional columns indicating each computation's place in
original list. The names of the addional columns are given by the \code{labels}
argument.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.data.frame}: Convert an errorcalc object to a data.frame
}}

