% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{aggregate_errors}
\alias{aggregate_errors}
\title{Aggregate a list of errors}
\usage{
aggregate_errors(x, ...)
}
\arguments{
\item{x}{an error object or a list of error objects}

\item{max_depth}{depth to check for an error object}
}
\value{
A single error object or a list of error objects
}
\description{
\code{aggregate_errors} combines two or more error objects
}
\details{
When only a single argument is provided, it should be a list
of error objects. When multiple arguments are provided, they
will be combined into a list using \code{list(x, ...)}.

In the simplest case where \code{x} is a list of error objects,
the function will return the error statistics of the combined
sample, which can be done by appropriately weighting the error
statistics themselves.
}
\author{
Matthew Zeigenfuse
}

