% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{data.frame-methods}
\alias{as.data.frame.error}
\alias{as.data.frame.error_list}
\alias{data.frame-methods}
\title{Create a data.frame}
\usage{
\method{as.data.frame}{error}(x, ...)

\method{as.data.frame}{error_list}(x, labels = NULL, label_name = "label",
  stringsAsFactors = default.stringsAsFactors(), ...)
}
\arguments{
\item{x}{an error or error_list object}

\item{...}{not implemented}

\item{labels}{labels to use instead of \code{names(x)}}

\item{label_name}{name of the label column}

\item{stringsAsFactors}{should strings be recoded as factors?}
}
\value{
a \code{data.frame}
}
\description{
\code{as.data.frame} converts an error or error_list object to a data.frame
}
\details{
For error_list objects, \code{as.data.frame} will be called for each list
element and the resulting \code{data.frame}s will be concatentated row-wise.
An addition column labelling the elements will be appended.
}

